requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
appdirs==1.4.4            # via black
appnope==0.1.2            # via ipykernel, ipython
argon2-cffi==20.1.0       # via notebook
argparse==1.4.0           # via -r requirements.in
astroid==2.4.2            # via pylint
async-generator==1.10     # via nbclient
attrs==20.3.0             # via jsonschema
backcall==0.2.0           # via ipython
black==20.8b1             # via -r requirements.in
bleach==3.2.1             # via nbconvert
boto3==1.16.40            # via -r requirements.in
botocore==1.19.40         # via boto3, s3transfer
certifi==2020.11.8        # via requests
cffi==1.14.4              # via argon2-cffi
chardet==3.0.4            # via requests
click==7.1.2              # via black
decorator==4.4.2          # via ipython
defusedxml==0.6.0         # via nbconvert
entrypoints==0.3          # via nbconvert
idna==2.10                # via requests
importlib-metadata==3.3.0  # via jsonschema
ipykernel==5.3.4          # via notebook
ipython-genutils==0.2.0   # via nbformat, notebook, traitlets
ipython==7.19.0           # via ipykernel
isort==5.6.4              # via pylint
jedi==0.17.2              # via ipython
jinja2==2.11.2            # via nbconvert, notebook
jmespath==0.10.0          # via boto3, botocore
jsonschema==3.2.0         # via nbformat
jupyter-client==6.1.7     # via ipykernel, nbclient, notebook
jupyter-core==4.7.0       # via jupyter-client, nbconvert, nbformat, notebook
jupyterlab-pygments==0.1.2  # via nbconvert
lazy-object-proxy==1.4.3  # via astroid
lxml==4.6.1               # via -r requirements.in
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via pylint
mistune==0.8.4            # via nbconvert
mypy-extensions==0.4.3    # via black
nbclient==0.5.1           # via nbconvert
nbconvert==6.0.7          # via notebook
nbformat==5.0.8           # via nbclient, nbconvert, notebook
nest-asyncio==1.4.3       # via nbclient
notebook==6.1.5           # via -r requirements.in
numpy==1.19.4             # via -r requirements.in, pandas
packaging==20.4           # via bleach
pandas==1.1.4             # via -r requirements.in
pandocfilters==1.4.3      # via nbconvert
parso==0.7.1              # via jedi
pathspec==0.8.1           # via black
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
prometheus-client==0.9.0  # via notebook
prompt-toolkit==3.0.8     # via ipython
ptyprocess==0.6.0         # via pexpect, terminado
pycparser==2.20           # via cffi
pygments==2.7.2           # via ipython, jupyterlab-pygments, nbconvert
pylint==2.6.0             # via -r requirements.in
pyparsing==2.4.7          # via packaging
pyrsistent==0.17.3        # via jsonschema
python-dateutil==2.8.1    # via botocore, jupyter-client, pandas
pytz==2020.4              # via pandas
pyzmq==20.0.0             # via jupyter-client, notebook
regex==2020.11.13         # via black
requests==2.25.0          # via -r requirements.in
ruamel.yaml.clib==0.2.2   # via ruamel.yaml
ruamel.yaml==0.16.12      # via -r requirements.in
s3transfer==0.3.3         # via boto3
send2trash==1.5.0         # via notebook
six==1.15.0               # via argon2-cffi, astroid, bleach, jsonschema, packaging, python-dateutil
terminado==0.9.1          # via notebook
testpath==0.4.4           # via nbconvert
toml==0.10.2              # via black, pylint
tornado==6.1              # via ipykernel, jupyter-client, notebook, terminado
traitlets==5.0.5          # via ipykernel, ipython, jupyter-client, jupyter-core, nbclient, nbconvert, nbformat, notebook
typed-ast==1.4.1          # via astroid, black
typing-extensions==3.7.4.3  # via black, importlib-metadata
urllib3==1.26.2           # via botocore, requests
wcwidth==0.2.5            # via prompt-toolkit
webencodings==0.5.1       # via bleach
wrapt==1.12.1             # via astroid
zipp==3.4.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
